+++ apps/integrations/src/generators/generateTypes.ts
@@ -1,0 +1,10 @@
+import { compile } from "json-schema-to-typescript";
+
+export async function getTypesFromSchema(schema: any, name: string) {
+  const ts = await compile(schema, name, {
+    additionalProperties: false,
+    bannerComment: "",
+    format: false,
+  });
+  return ts;
+}