+++ apps/integrations/src/integrations/slack/index.ts
@@ -1,0 +1,13 @@
+import { Service } from "core/service/types";
+import { authentication } from "./authentication";
+import * as actions from "./actions/actions";
+
+export const slackv2: Service = {
+  name: "Slack",
+  service: "slackv2",
+  version: "2.0.0",
+  live: false,
+  authentication,
+  actions,
+  retryableStatusCodes: [408, 429, 500, 502, 503, 504],
+};